refactor FloatingActionButton ref use, and renam the file to index.js
